Tweaked notifications to alert about success or failure uploading. Fixed bug that prevented settings from carrying over correctly (hopefully).

+++ OctoprintUploader.xaml.cs
@@ -16,6 +16,7 @@ using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
 
+
 namespace OctoUploader
 {
     /// <summary>
@@ -71,7 +72,7 @@ namespace OctoUploader
             {
                 rkApp.SetValue(System.Diagnostics.Process.GetCurrentProcess().ProcessName, System.Reflection.Assembly.GetExecutingAssembly().Location);
             }
-            else
+            else if ( rkApp.GetValue (System.Diagnostics.Process.GetCurrentProcess().ProcessName) != null  )
             {
                 rkApp.DeleteValue(System.Diagnostics.Process.GetCurrentProcess().ProcessName);
             }
